[mod] #102 걸음 수 커스텀 뷰
