+++ motivoo_design/src/main/res/layout/motivoo_step_count_text.xml
@@ -53,7 +53,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toEndOf="@id/view_line_divider"
         app:layout_constraintTop_toTopOf="parent"
-        tools:text="부모님 걸음"/>
+        tools:text="부모님 걸음" />
 
     <TextView
         android:id="@+id/tv_other_step_count"
